requirements.txt
# ============================================================
# Armenian Voice AI Support Agent — Requirements
# ============================================================

# --- Scraping ---
playwright>=1.44.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
cloudscraper>=1.2.71
requests>=2.31.0

# --- RAG ---
chromadb>=0.5.0
openai>=1.30.0
python-dotenv>=1.0.0

# --- LiveKit Agent ---
livekit-agents>=0.8.0
livekit-plugins-openai>=0.8.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-elevenlabs>=0.6.0
livekit-plugins-silero>=0.6.0     # VAD (Voice Activity Detection)